chore: increased id capacity from int32 to int64

+++ app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, func
+from sqlalchemy import BigInteger, Column, Integer, String, DateTime, func
 from .database import Base
 
 
@@ -6,6 +6,6 @@ class User(Base):
     __tablename__ = 'users'
 
     id = Column(Integer, primary_key=True, index=True)
-    chat_id = Column(Integer, unique=True, index=True)
+    chat_id = Column(BigInteger, unique=True, index=True)
     username = Column(String, index=True, nullable=True)
     subscribe_date = Column(DateTime(timezone=True), server_default=func.now())
